Add certification to edit/remove mondais

+++ socket.js
@@ -148,11 +148,17 @@ module.exports = function(socket) {
   socket.on('removeMondaiChat', function(data) {});
 
   socket.on('mondaiMessage', function(msg) {
+    if(mondai[socket.room])
+      if(mondai[socket.room].removePass !== msg.removePass){
+        console.log('Invalid removepass');
+        return;
+      } 
     var doc = {
       room: socket.room,
       sender: socket.name,
-      content: String(msg.content || 'クリックして問題文を入力'),
-      trueAns: trueAns[socket.room] || 'クリックして解説を入力',
+      removePass: msg.removePass,
+      content: String(msg.content || '問題文'),
+      trueAns: String(trueAns[socket.room] || '解説'),
       created_month: msg.created_month.toString(),
       created_date: msg.created_date.toString()
     };
@@ -276,8 +282,17 @@ module.exports = function(socket) {
         .emit('lobbyChat', reverseById(lobbyChats));
     });
   });
-  socket.on('clear', function() {
+  socket.on('clear', function(removePass) {
     var room = socket.room;
+    if(!mondai[room]){
+      console.log('no room');
+      return;
+    }
+    else
+      if(mondai[room].removePass !== removePass){
+        console.log('invalid removepass');
+        return;
+      }
     mondai[room] = null;
     trueAns[room] = null;
     client.del(room);
